+++ lib/commercetools-api/src/Models/Cart/CartDraft.php
@@ -45,6 +45,7 @@ interface CartDraft extends JsonObject
     public const FIELD_SHIPPING = 'shipping';
     public const FIELD_ITEM_SHIPPING_ADDRESSES = 'itemShippingAddresses';
     public const FIELD_DISCOUNT_CODES = 'discountCodes';
+    public const FIELD_DIRECT_DISCOUNTS_IGNORE_CART_DISCOUNTS = 'directDiscountsIgnoreCartDiscounts';
     public const FIELD_COUNTRY = 'country';
     public const FIELD_LOCALE = 'locale';
     public const FIELD_ORIGIN = 'origin';
@@ -267,6 +268,17 @@ interface CartDraft extends JsonObject
     public function getDiscountCodes();
 
     /**
+     * <ul>
+     * <li>If set to <code>true</code>, only <a href="ctp:api:type:DirectDiscount">Direct Discounts</a> apply to the Cart. Matching <a href="ctp:api:type:CartDiscount">Cart Discounts</a> are ignored, and Discount Codes cannot be added.</li>
+     * <li>If set to <code>false</code>, Cart Discounts, Discount Codes, and Direct Discounts apply to the Cart.</li>
+     * </ul>
+     *
+
+     * @return null|bool
+     */
+    public function getDirectDiscountsIgnoreCartDiscounts();
+
+    /**
      * <p>Used for <a href="/api/pricing-and-discounts-overview#line-item-price-selection">Line Item price selection</a>.
      * If used for <a href="ctp:api:endpoint:/{projectKey}/in-store/carts:POST">Create Cart in Store</a>, the provided country must be one of the <a href="ctp:api:type:Store">Store's</a> <code>countries</code>.</p>
      *
@@ -442,6 +454,11 @@ interface CartDraft extends JsonObject
      * @param ?array $discountCodes
      */
     public function setDiscountCodes(?array $discountCodes): void;
+
+    /**
+     * @param ?bool $directDiscountsIgnoreCartDiscounts
+     */
+    public function setDirectDiscountsIgnoreCartDiscounts(?bool $directDiscountsIgnoreCartDiscounts): void;
 
     /**
      * @param ?string $country